requirements.txt
pyspark
requests
django
djangorestframework